Refactored file structure fo easier follow through
